+++ packages/apollo/test/components/Mutation.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import gql from 'graphql-tag';
 import { WrappingComponent } from '@airbnb/lunar-test-utils';
-import Loader from '@airbnb/lunar/lib/components/Loader';
+// import Loader from '@airbnb/lunar/lib/components/Loader';
 import ErrorMessage from '@airbnb/lunar/lib/components/ErrorMessage';
 import { MockedProvider } from 'react-apollo/test-utils';
 import Mutation from '../../src/components/Mutation';
@@ -47,19 +47,19 @@ describe('Mutation', () => {
       },
     };
 
-    it('renders a `Loader` by default', () => {
-      const wrapper = renderer.create(
-        <MockedProvider mocks={[mock]} addTypename={false}>
-          <WrappingComponent>
-            <Mutation mutation={MUTATION}>{childHandler}</Mutation>
-          </WrappingComponent>
-        </MockedProvider>,
-      );
+    // it('renders a `Loader` by default', () => {
+    //   const wrapper = renderer.create(
+    //     <MockedProvider mocks={[mock]} addTypename={false}>
+    //       <WrappingComponent>
+    //         <Mutation mutation={MUTATION}>{childHandler}</Mutation>
+    //       </WrappingComponent>
+    //     </MockedProvider>,
+    //   );
 
-      wrapper.root.findByType('button').props.onClick();
+    //   wrapper.root.findByType('button').props.onClick();
 
-      expect(wrapper.root.findByType(Loader)).toBeDefined();
-    });
+    //   expect(wrapper.root.findByType(Loader)).toBeDefined();
+    // });
 
     it('can pass a custom loading element with `loading` prop', () => {
       const loader = <div>Loading!</div>;